requirements.txt
selenium
webdriver-manager
fastapi
uvicorn[standard]
questionary
rich
openpyxl
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
